+++ tests/test_rules/grape.rb
@@ -14,10 +14,9 @@ module Twitter
       end
     end
 
-    # todoruleid: grape-route-authenticated
     resource :statuses do
       desc 'Return a public timeline.'
-      # todoruleid: grape-route-authenticated
+      # ruleid: grape-route-unauthenticated
       get :public_timeline do
         Status.limit(20)
       end
@@ -29,12 +28,27 @@ module Twitter
         current_user.statuses.limit(20)
       end
 
+      desc 'Return a personal timeline.'
+      # ruleid: grape-route-authenticated
+      get :home_timeline do
+        authenticate!
+        authorize!
+        current_user.statuses.limit(20)
+      end
+
+      desc 'Return a personal timeline.'
+      # ruleid: grape-route-authorized
+      get :home_timeline do
+        authorize!
+        current_user.statuses.limit(20)
+      end
+
       desc 'Return a status.'
       params do
         requires :id, type: Integer, desc: 'Status ID.'
       end
       route_param :id do
-        # todoruleid: grape-route-authenticated
+        # ruleid: grape-route-unauthenticated
         get do
           Status.find(params[:id])
         end
@@ -76,6 +90,50 @@ module Twitter
         authenticate!
         current_user.statuses.find(params[:id]).destroy
       end
+    end
+  end
+
+  class API < Grape::API
+    helpers do
+      def current_user
+        User.find(params[:user_id])
+      end
+
+      def authenticate!
+        error!('401 Unauthorized', 401) unless current_user
+      end
+    end
+
+    helpers StatusHelpers, HttpCodesHelpers
+
+    before do
+      authenticate!
+    end
+
+    # ruleid: grape-route-authenticated
+    get 'info' do
+      user_info(current_user)
+    end
+  end
+
+  class API < Grape::API
+    helpers do
+      def current_user
+        User.find(params[:user_id])
+      end
+
+      def authenticate!
+        error!('401 Unauthorized', 401) unless current_user
+      end
+    end
+
+    helpers StatusHelpers, HttpCodesHelpers
+
+    before { authenticate! }
+
+    # ruleid: grape-route-authenticated
+    get 'info' do
+      user_info(current_user)
     end
   end
 end